updated model index.js, updated include section of all routes

+++ routes/api/product-routes.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const { Product, Category: Product, Tag, ProductTag } = require('../../models');
+const { Product, Category, Tag, ProductTag } = require('../../models');
 
 // The `/api/products` endpoint
 
@@ -9,7 +9,9 @@ router.get('/', async (req, res) => {
   // be sure to include its associated Category and Tag data
   try {
     const productData = await Product.findAll({
-      // include: [{ }]
+      include: [{ model: Category, attributes: ['category_name'] }, {
+        model: Tag, through: ProductTag
+      }]
     });
     res.status(200).json(productData);
   } catch (err) {
@@ -23,7 +25,9 @@ router.get('/:id', async (req, res) => {
   // be sure to include its associated Category and Tag data
   try {
     const productData = await Product.findByPk(req.params.id, {
-      // include: [{ }]
+      include: [{ model: Category, attributes: ['category_name'] }, {
+        model: Tag, through: ProductTag
+      }]
     });
     if (!productData) {
       res.status(404).json({ message: "No Product found with this ID" });
@@ -41,7 +45,8 @@ router.post('/', (req, res) => {
       product_name: "Basketball",
       price: 200.00,
       stock: 3,
-      tagIds: [1, 2, 3, 4]
+      tagIds: [1, 2, 3, 4],
+      category_id: 1
     }
   */
   Product.create(req.body)
@@ -66,7 +71,48 @@ router.post('/', (req, res) => {
     });
 });
 
-// update product
+// // update product
+// router.put('/:id', (req, res) => {
+//   // update product data
+//   Product.update(req.body, {
+//     where: {
+//       id: req.params.id,
+//     },
+//   })
+//     .then((product) => {
+//       // find all associated tags from ProductTag
+//       return ProductTag.findAll({ where: { product_id: req.params.id } });
+//     })
+//     .then((productTags) => {
+//       // get list of current tag_ids
+//       const productTagIds = productTags.map(({ tag_id }) => tag_id);
+//       // create filtered list of new tag_ids
+//       const newProductTags = req.body.tagIds
+//         .filter((tag_id) => !productTagIds.includes(tag_id))
+//         .map((tag_id) => {
+//           return {
+//             product_id: req.params.id,
+//             tag_id,
+//           };
+//         });
+//       // figure out which ones to remove
+//       const productTagsToRemove = productTags
+//         .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
+//         .map(({ id }) => id);
+
+//       // run both actions
+//       return Promise.all([
+//         ProductTag.destroy({ where: { id: productTagsToRemove } }),
+//         ProductTag.bulkCreate(newProductTags),
+//       ]);
+//     })
+//     .then((updatedProductTags) => res.json(updatedProductTags))
+//     .catch((err) => {
+//       // console.log(err);
+//       res.status(400).json(err);
+//     });
+// });
+
 router.put('/:id', (req, res) => {
   // update product data
   Product.update(req.body, {
@@ -75,38 +121,53 @@ router.put('/:id', (req, res) => {
     },
   })
     .then((product) => {
-      // find all associated tags from ProductTag
-      return ProductTag.findAll({ where: { product_id: req.params.id } });
-    })
-    .then((productTags) => {
-      // get list of current tag_ids
-      const productTagIds = productTags.map(({ tag_id }) => tag_id);
-      // create filtered list of new tag_ids
-      const newProductTags = req.body.tagIds
-        .filter((tag_id) => !productTagIds.includes(tag_id))
-        .map((tag_id) => {
-          return {
-            product_id: req.params.id,
-            tag_id,
-          };
+      if (req.body.tagIds && req.body.tagIds.length) {
+        const productTags = ProductTag.findAll({
+          where: { product_id: req.params.id }
         });
-      // figure out which ones to remove
-      const productTagsToRemove = productTags
-        .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
-        .map(({ id }) => id);
-
-      // run both actions
-      return Promise.all([
-        ProductTag.destroy({ where: { id: productTagsToRemove } }),
-        ProductTag.bulkCreate(newProductTags),
-      ]);
+        const productTagIds = productTags.map(({ tag_id }) => tag_id);
+        // create filtered list of new tag_ids
+        const newProductTags = req.body.tagIds
+          .filter((tag_id) => !productTagIds.includes(tag_id))
+          .map((tag_id) => {
+            return {
+              product_id: req.params.id,
+              tag_id,
+            };
+          });
+        // figure out which ones to remove
+        const productTagsToRemove = productTags
+          .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
+          .map(({ id }) => id);
+          return Promise.all([
+            ProductTag.destroy({ where: { id: productTagsToRemove } }),
+            ProductTag.bulkCreate(newProductTags),
+          ]);
+        }
+        return res.json("success");
+      })
+      .catch((err) => {
+        // console.log(err);
+        res.status(400).json(err);
+      });
+  });
+  
+  router.delete('/:id', (req, res) => {
+    Product.destroy({
+      where: {
+        id: req.params.id,
+      },
     })
-    .then((updatedProductTags) => res.json(updatedProductTags))
-    .catch((err) => {
-      // console.log(err);
-      res.status(400).json(err);
-    });
-});
+      .then((products) => {
+        console.log(products);
+        res.json(products);
+      })
+      .catch((err) => {
+        console.log(err);
+        res.status(400).json(err);
+      });
+  });
+  
 
 router.delete('/:id', async (req, res) => {
   // delete one product by its `id` value
